+++ post_office/settings.py
@@ -3,16 +3,7 @@ import warnings
 from django.conf import settings
 from django.core.cache.backends.base import InvalidCacheBackendError
 
-from .compat import import_attribute
-
-
-try:
-    from django.core.cache import caches
-
-    def get_cache(name):
-        return caches[name]
-except ImportError:
-    from django.core.cache import get_cache
+from .compat import import_attribute, get_cache
 
 
 def get_backend(alias='default'):