+++ post_office/tests/test_models.py
@@ -102,7 +102,7 @@ class ModelTest(TestCase):
         Ensure that status and log are set properly on sending failure
         """
         email = Email.objects.create(to=['to@example.com'], subject='Test',
-                                     from_email='from@example.com', 
+                                     from_email='from@example.com',
                                      message='Message', backend_alias='random')
         # Ensure that after dispatch status and logs are correctly set
         email.dispatch()
@@ -208,12 +208,14 @@ class ModelTest(TestCase):
         form = EmailTemplateForm(data)
 
         self.assertFalse(form.is_valid())
-        self.assertEqual(form.errors, {
-            'default_template': [u'This field is required.'],
-            'content': [u"Invalid filter: 'titl'"],
-            'html_content': [u'Unclosed tags: endblock '],
-            'subject': [u'Empty variable tag']
-        })
+
+        self.assertEqual(form.errors['default_template'],  [u'This field is required.'])
+        self.assertEqual(form.errors['content'], [u"Invalid filter: 'titl'"])
+        self.assertIn(form.errors['html_content'],
+                      [[u'Unclosed tags: endblock '],
+                       [u"Unclosed tag on line 1: 'block'. Looking for one of: endblock."]])
+        self.assertIn(form.errors['subject'],
+                      [[u'Empty variable tag'], [u'Empty variable tag on line 1']])
 
     def test_string_priority(self):
         """
@@ -238,6 +240,14 @@ class ModelTest(TestCase):
             'Invalid priority, must be one of: low, medium, high, now'
         )
 
+    def test_send_recipient_display_name(self):
+        """
+        Regression test for:
+        https://github.com/ui/django-post_office/issues/73
+        """
+        email = send(recipients=['Alice Bob <email@example.com>'], sender='from@a.com')
+        self.assertTrue(email.to)
+
     def test_attachment_filename(self):
         attachment = Attachment()
 
@@ -246,7 +256,7 @@ class ModelTest(TestCase):
             content=ContentFile('test file content'),
             save=True
         )
-        self.assertEquals(attachment.name, 'test.txt')
+        self.assertEqual(attachment.name, 'test.txt')
 
     def test_attachments_email_message(self):
         email = Email.objects.create(to=['to@example.com'],
@@ -267,3 +277,9 @@ class ModelTest(TestCase):
         template = EmailTemplate.objects.create(name='name')
         id_template = template.translated_templates.create(language='id')
         self.assertEqual(id_template.name, template.name)
+
+    def test_models_repr(self):
+        self.assertEqual(repr(EmailTemplate(name='test', language='en')),
+                         '<EmailTemplate: test en>')
+        self.assertEqual(repr(Email(to=['test@example.com'])),
+                         "<Email: ['test@example.com']>")